+++ src/main/java/tc/TrajectorySlotMapper.java
@@ -5,7 +5,7 @@ import org.apache.spark.api.java.function.PairFunction;
 import java.io.Serializable;
 import scala.Tuple2;
 
-public class TrajectorySlotMapper implements PairFunction<String, Integer, TCPoint>, Serializable {
+public class TrajectorySlotMapper implements PairFunction<String, Long, TCPoint>, Serializable {
 
     private int _slotInterval = 0;
 
@@ -15,7 +15,7 @@ public class TrajectorySlotMapper implements PairFunction<String, Integer, TCPoi
     }
 
     @Override
-    public Tuple2<Integer, TCPoint> call(String line) throws Exception {
+    public Tuple2<Long, TCPoint> call(String line) throws Exception {
         String[] split = line.split(",");
         Integer objectId = Integer.parseInt(split[0]);
         Double x = Double.parseDouble(split[1]);
@@ -30,7 +30,7 @@ public class TrajectorySlotMapper implements PairFunction<String, Integer, TCPoi
             timestamp = Integer.parseInt(split[3]);
         }
         TCPoint point = new TCPoint(objectId, x, y, timestamp);
-        int slotId = (int)Math.ceil(timestamp / _slotInterval);
+        long slotId = (int)Math.ceil(timestamp / _slotInterval);
 
         return new Tuple2<>(slotId, point);
     }
